+++ to.etc.domui/src/main/java/to/etc/domui/converter/MsDurationConverter.java
@@ -89,7 +89,7 @@ public class MsDurationConverter implements IConverter<Long> {
 						dur += nr;
 						ms.accept();
 					} else {
-						dur += nr * 60 * 1000;
+						dur += (long) nr * 60 * 1000;
 					}
 					break;
